refactor(context): move certain functions
from useOnboardingFlow to useOnboardingVisibility

functions dont control flow, they control visibility
